# swagger.yaml
openapi: 3.0.0
info:
  title: vinyl-shop
  description: "The app represents an API of a mock vinyl shop.  \nSome endpoints are accessible to all users (e.g. when browsing the inventory).  \n\nUsers can register and login to have access to the ordering and other functionalities - userAuthentication middleware enables this.  \nAdmin permissions can be generated by accessing the DB directly only in order to avoid regular users assigning themselves privileges they aren't supposed to have. Admin users can create new bands/labels/styles/users/inventory items and so on, edit and delete existing ones and basically manipulate the DB.  \n\nEach vinyl must be linked to an existing label and band (meaning these need to be created in advance).  \n\nThe ordering process works as follow:  \nAn INSERT query INTO a cart intermediary table is generated since the link between album and user is many-2-many, e.g. one user can order multiple albums, one album could be ordered by multiple users. Each cart (naming is not ideal) will have an unique ID (which we are not really interested in)  and a cart_number which will be our unique identifier as far as an actual cart containing multiple items is being implemented. When the user places an order we insert into purchases an entry related to the the given cart_no - hence we know which user has ordered what albums in an order with this cart_no."
  version: '1.0'
servers:
- url: http://localhost:3000/api/v1
  variables: {}
...
  variables: {}
paths:
...
  /users:
    get:
      tags:
      - users
      summary: GET all USERS
      operationId: GETallUSERS
      parameters: []
      responses:
        '200':
          description: successful operation
      deprecated: false
      security:
      - bearer: []
  /users/{userId}:
    get:
      tags:
      - users
      summary: GET USER by ID
      operationId: GETUSERbyID
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: ID of the user
      responses:
        '200':
          description: successful operation
        '404':
          description: not found
        '400':
          description: bad request
        '401':
          description: unauthorized
      deprecated: false
      security:
      - bearer: []
      
    put:
      tags:
      - users
      summary: UPDATE USER
      operationId: UPDATEUSER
      parameters: 
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: ID of the user
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UPDATEUSERREQUEST'
              - example:
                  f_name: bbbb
                  l_name: bbbb
                  email: bb@bb.bb
                  password: bbbb
                  house_number: 22
                  street_name: second street
                  city: London
                  country: France
                  is_admin: true
                  is_contributor: true
            example:
              f_name: bbbb
              l_name: bbbb
              email: bb@bb.bb
              password: bbbb
              house_number: 22
              street_name: second street
              city: London
              country: France
              is_admin: true
              is_contributor: true
        required: true
      responses:
        '201':
          description: successful operation
        '404':
          description: not found
        '400':
          description: bad request
        '401':
          description: unauthorized
      deprecated: false
      security:
      - bearer: []

    delete:
      tags:
      - users
      summary: DELETE USER
      operationId: DELETEUSER
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: ID of the user
      responses:
        '204':
          description: successful operation
        '404':
          description: not found
        '400':
          description: bad request
        '401':
          description: unauthorized
      deprecated: false
      security:
      - bearer: []
...
components:
  schemas:
...
    UPDATEUSERREQUEST:
      title: UPDATEUSERREQUEST
      required:
      - f_name
      - l_name
      - email
      - password
      - house_number
      - street_name
      - city
      - country
      - is_admin
      - is_contributor
      type: object
      properties:
        f_name:
          type: string
        l_name:
          type: string
        email:
          type: string
        password:
          type: string
        house_number:
          type: integer
          format: int32
        street_name:
          type: string
        city:
          type: string
        country:
          type: string
        is_admin:
          type: boolean
        is_contributor:
          type: boolean
      example:
        f_name: bbbb
        l_name: bbbb
        email: bb@bb.bb
        password: bbbb
        house_number: 22
        street_name: second street
        city: London
        country: France
        is_admin: true
        is_contributor: true
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...